# backend/services/question_service/docs/api/OPENAPI_EG_DOCS.yaml
openapi: 3.0.3
info:
  title: Question Service API (PeerPrep)
  version: 0.3.0
  description: |
    Microservice to manage and serve coding interview questions.

    Key behaviors:
    - Questions have:
      - problem statement in Markdown (`body_md`)
      - sanitized HTML (`body_html`) with signed image URLs on read
      - attachments stored in S3, surfaced through presigned/signed URLs
      - difficulty, topics, version, status
      - execution resources:
        - starter code (currently python only)
        - test cases (sample vs hidden)

    - Topics are managed separately via `/admin/topics` and referenced by slug.

    - `/select` is idempotent for a `matching_id` for ~10 minutes.

    - Admin endpoints require `admin` role (via `x-role: admin` in dev or RBAC'ed JWT in prod).

    - Service-to-service calls (like the matching worker) may use `x-role: service`.
servers:
  - url: http://localhost:3000
    description: Local dev
  - url: https://api.example.com/question-service
    description: Production (placeholder)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Production calls should include a Bearer token.
        In local/dev we also allow `x-role: admin` or `x-role: service`.
...
  schemas:
    Difficulty:
      type: string
      enum: [Easy, Medium, Hard]
    Status:
      type: string
      enum: [draft, published, archived]
    Error:
      type: object
      properties:
        error:
          type: string
          example: not_found
        message:
          type: string
          example: Question not found
      required:
        - error
    Attachment:
      type: object
      description: Attachment metadata stored on the question row.
      properties:
        object_key:
          type: string
          example: questions/two-sum/20251031/abc123-diagram.png
        mime:
          type: string
          example: image/png
        alt:
          type: string
          nullable: true
          example: diagram
        byte_size:
          type: integer
          minimum: 0
          nullable: true
        width:
          type: integer
          minimum: 1
          nullable: true
        height:
          type: integer
          minimum: 1
          nullable: true
      required:
        - object_key
        - mime
    TopicTag:
      type: object
      description: Topic metadata (used in read models).
      additionalProperties: false
      properties:
        slug:
          type: string
          example: arrays
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
        color_hex:
          type: string
          example: '#3b82f6'
          pattern: '^#(?:[0-9a-fA-F]{6}|[0-9a-fA-F]{3})$'
      required:
        - slug
        - color_hex
...
    TestCaseWrite:
      type: object
      description: |
        Test case provided by admin on create/patch.
      properties:
        visibility:
          type: string
          enum: [sample, hidden]
          description: |
            - sample -> visible to users
            - hidden -> internal only, not exposed publicly
        input_data:
          type: string
          description: Raw input (JSON, stdin args, etc.)
        expected_output:
          type: string
          description: Expected output (JSON, stdout, etc.)
        ordinal:
          type: integer
          description: Stable order for runner / UI
          example: 0
      required:
        - visibility
        - input_data
        - expected_output
    TestCaseAdminRead:
      type: object
      description: Test case returned to admin/service.
      properties:
        name:
          type: string
          example: case-0
        visibility:
          type: string
          enum: [sample, hidden]
        input:
          type: string
          example: '[[1,2],[3,4]]'
        expected:
          type: string
          example: '[[3,1],[4,2]]'
        ordinal:
          type: integer
          example: 0
      required:
        - visibility
        - input
        - expected
        - ordinal
...
    QuestionBase:
      type: object
      properties:
        id:
          type: string
          example: rotate-image-in-place
        title:
          type: string
        body_md:
          type: string
          description: |
            Markdown body.

            Inline images in authoring can refer to internal objects using:
            `pp://<object_key>`.

            These keys are finalized on create/patch:
            `staging/...` → `questions/{slug}/...`.
        difficulty:
          $ref: '#/components/schemas/Difficulty'
        topics:
          type: array
          description: |
            Topics for this question.

            On READ (admin/public):
            - array of `{ slug, color_hex }`.

            On WRITE (create/patch):
            - just an array of slugs.
          items:
            $ref: '#/components/schemas/TopicTag'
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
        status:
          $ref: '#/components/schemas/Status'
        version:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - title
        - body_md
        - difficulty
        - status
        - version
...
    QuestionUpdate:
      type: object
      description: |
        Partial update to an existing draft.

        Behavior:
        - Any provided simple fields (title, difficulty, etc.) are updated.
        - If `attachments` includes new `staging/...` keys, those are finalized under
          `questions/{id}/...`, and if `body_md` is also provided, its `pp://staging/...`
          refs are rewritten to the finalized keys.
        - If `starter_code_python` is provided, the python starter row is upserted.
        - If `test_cases` is provided, ALL existing test cases are replaced by the new list.
        - If `topics` is provided, the entire topic set is replaced.
          All slugs must already exist in `/admin/topics`.
      properties:
        title:
          type: string
        body_md:
          type: string
        difficulty:
          $ref: '#/components/schemas/Difficulty'
        topics:
          type: array
          description: |
            Array of topic slugs. All slugs must already exist in `/admin/topics`.
          items:
            type: string
          example:
            - arrays
            - matrix
        attachments:
          type: array
          items:
            type: object
            required:
              - object_key
              - mime
            properties:
              object_key:
                type: string
              mime:
                type: string
              alt:
                type: string
                nullable: true
        starter_code_python:
          type: string
          example: |
            def rotate_image(matrix):
                # improved version
                pass
        test_cases:
          type: array
          items:
            $ref: '#/components/schemas/TestCaseWrite'
    QuestionAdminRead:
      allOf:
        - $ref: '#/components/schemas/QuestionBase'
        - type: object
          description: |
            Admin/internal full view (works for any status):
            - question metadata
            - sanitized body_html with signed URLs
            - ALL test cases (including hidden)
            - starter_code_python
          properties:
            body_html:
              type: string
              description: Sanitized HTML derived from body_md, with signed image URLs.
            starter_code_python:
              type: string
              nullable: true
              description: Pulled from `question_python_starter`.
            test_cases:
              type: array
              description: All test cases, including `hidden`.
              items:
                $ref: '#/components/schemas/TestCaseAdminRead'
          required:
            - body_html
...
security:
  - bearerAuth: []
...
paths:
...
  /admin/questions/{id}:
    get:
      tags:
        - Admin
      summary: Get full question (draft / published / archived)
      description: |
        Returns the internal/admin bundle:
        - question metadata
        - sanitized `body_html` with signed URLs
        - ALL test cases (including hidden)
        - starter_code_python
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionAdminRead'
        '401':
          description: Unauthorized
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []
      x-roles:
        - admin

    patch:
      tags:
        - Admin
      summary: Update an existing draft
      description: |
        Partial update.

        Behavior:
        - Any provided simple fields (title, difficulty, etc.) are updated.
        - If `attachments` includes new `staging/...` keys, those are finalized
          under `questions/{id}/...`, and if `body_md` is provided,
          its `pp://staging/...` refs are rewritten.
        - If `starter_code_python` is provided, python starter is upserted.
        - If `test_cases` is provided, ALL test cases are replaced by the new list.
        - If `topics` is provided, the entire topic set is replaced
          (all slugs must already exist via `/admin/topics`).

        Returns the updated admin view, including hidden tests.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionUpdate'
      responses:
        '200':
          description: Updated draft
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionAdminRead'
        '400':
          description: Validation error (e.g. unknown_topics)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
      security:
        - bearerAuth: []
      x-roles:
        - admin

    delete:
      tags:
        - Admin
      summary: Archive a published question
      description: |
        Soft-delete semantics:
        - Only allowed if current status is `published`.
        - Updates `questions.status = archived`,
          bumps `version`, writes snapshot.
        - Returns the updated row as JSON.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Archived
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionAdminRead'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found or not eligible
      security:
        - bearerAuth: []
      x-roles:
        - admin
...